Working on jump. Added ships back to game_data. Made the start name Unknown and added a start ship. Stable.

+++ cogs/system.py
@@ -1,5 +1,6 @@
 from cogs.utils.embed import send_error_embed
 from cogs.utils.time import now
+from config import jdata
 from discord.ext import commands
 from replit import db
 
@@ -27,6 +28,17 @@ class System(commands.Cog):
             return
 
         user_dlg_id = db['users'][user_id]['dialogue_id']
+
+        print(f'{now()}: Checking if "{self.jump.name}" command is unlocked.')
+        unlocked = True
+        for tag in ['major', 'minor', 'sub']:
+            if int(user_dlg_id[tag]) < int(jdata['game_data']['commands'][self.jump.name]['unlock']['dialogue'][tag]):
+                unlocked = False
+                
+        print(f'{now()}: "{self.jump.name}" command is {"unlocked" if unlocked else "locked"}.')
+        if not unlocked:
+            await send_error_embed(ctx, 'command_locked')
+            return
         
         await ctx.message.delete()
         await ctx.send(f'Pong: {round(self.bot.latency * 1000)}ms.')
